refactor: extract reusable components
